+++ fk_str.h
@@ -2,7 +2,7 @@
 #define _FK_STR_H_
 
 typedef struct _fk_str {
-	int len;/*full length, including '\0' */
+	size_t len;/*full length, including '\0' */
 	char seq[];
 } fk_str;
 
@@ -10,7 +10,7 @@ typedef struct _fk_str {
 
 #define fk_str_len(str) ((str)->len)
 
-fk_str *fk_str_create(char *seq, int len);
+fk_str *fk_str_create(char *seq, size_t len);
 void fk_str_destroy(fk_str *src);
 fk_str *fk_str_clone(fk_str *src);
 int fk_str_cmp(fk_str *s1, fk_str *s2);
